update logic onMount was moved to SearchSection
